refactor: timer utils
